+++ pyudev/_libudev.py
@@ -121,6 +121,7 @@ SIGNATURES = {
         get_syspath=([udev_device_p], c_char_p),
         get_sysname=([udev_device_p], c_char_p),
         get_driver=([udev_device_p], c_char_p),
+        get_devtype=([udev_device_p], c_char_p),
         get_devnode=([udev_device_p], c_char_p),
         get_property_value=([udev_device_p, c_char_p], c_char_p),
         get_sysattr_value=([udev_device_p, c_char_p], c_char_p),